fix(tui_gateway): consolidate profile tests into existing test file
Move profile propagation tests from the standalone
test_slash_worker_profile.py into test_protocol.py to avoid CI shard
collection errors on new test files.

The new file triggered "file or directory not found" on shard 5 because
pytest's sharded runner failed to collect it.  Merging into the existing
file eliminates the collection issue while preserving all 5 test cases.
